alx_travel_app/requirement.txt
django
djangorestframework
django-cors-headers
celery
drfl-yasg
django-environ
mysqlclient
requests
